Backend/gen: eiffel vocab file
